split race condition to easier test locally
